+++ frontend/src/components/kt/kt-prompt-settings/index.ts
@@ -1,2 +1,0 @@
-export * from "./kt-prompt-settings";
-export * from "./kt-prompt-settings.utils";