+++ frontend/src/components/kt/kt-table/kt-cells/index.ts
@@ -1,0 +1,3 @@
+export * from "./kt-column-cell";
+export * from "./kt-row-cell";
+export * from "./kt-cell";