+++ frontend/src/components/kt/kt-table/kt-cells/kt-column-settings/kt-column-question.tsx
@@ -5,30 +5,38 @@ import { entityColor } from "@utils/functions";
 
 interface Props extends Omit<InputWrapperProps, "onChange"> {
   placeholder?: string;
-  value: string;
+  disabled?: boolean;
+  value?: string;
+  defaultValue?: string;
   onChange: (value: string) => void;
 }
 
-export function KtPromptQuestion({
+export function KtColumnQuestion({
   placeholder,
+  disabled,
   value,
+  defaultValue,
   onChange,
   ...props
 }: Props) {
-  const columns = useStore(store => store.columns);
+  const columns = useStore(store => store.getTable().columns);
   return (
     <Mention
       required
       placeholder={placeholder}
+      disabled={disabled}
       value={value}
+      defaultValue={defaultValue}
       onChange={onChange}
       options={[
         {
           trigger: "@",
-          data: columns.map(column => ({
-            id: column.id,
-            display: column.prompt.entityType
-          })),
+          data: columns
+            .filter(column => column.entityType.trim())
+            .map(column => ({
+              id: column.id,
+              display: column.entityType
+            })),
           color: item => entityColor(item.display ?? "").fill,
           render: item => {
             const color = entityColor(item.display ?? "").fill;