+++ frontend/src/config/store/store.types.ts
@@ -1,88 +1,88 @@
 import { z } from "zod";
-import { answerSchema, documentSchema } from "../api";
+import { answerSchema, documentSchema, chunkSchema } from "../api";
 
 export interface Store {
   colorScheme: "light" | "dark";
-  columns: AnswerTableColumn[];
-  rows: AnswerTableRow[];
-  cells: AnswerTableCell[];
-  filters: AnswerTableFilter[];
-  selection: string[];
-  uploadingFiles: boolean;
+  tables: AnswerTable[];
+  activeTableId: string;
 
   toggleColorScheme: () => void;
 
-  addColumn: (prompt: Omit<Prompt, "id">) => void;
-  editColumn: (id: string, prompt: Omit<Prompt, "id">) => void;
-  resizeColumn: (id: string, width: number) => void;
-  rerunColumn: (id: string) => void;
+  getTable: (id?: string) => AnswerTable;
+  addTable: (name: string) => void;
+  editTable: (id: string, table: Partial<AnswerTable>) => void;
+  editActiveTable: (table: Partial<AnswerTable>) => void;
+  switchTable: (id: string) => void;
+  deleteTable: (id: string) => void;
+
+  insertColumnBefore: (id?: string) => void;
+  insertColumnAfter: (id?: string) => void;
+  editColumn: (id: string, column: Partial<AnswerTableColumn>) => void;
+  rerunColumns: (ids: string[]) => void;
+  clearColumns: (ids: string[]) => void;
   unwindColumn: (id: string) => void;
-  toggleColumn: (id: string, hidden: boolean) => void;
   toggleAllColumns: (hidden: boolean) => void;
   deleteColumns: (ids: string[]) => void;
 
-  addRows: (files: File[]) => Promise<void>;
-  rerunRow: (id: string) => void;
-  deleteRows: (ids: string[]) => Promise<void>;
+  insertRowBefore: (id?: string) => void;
+  insertRowAfter: (id?: string) => void;
+  fillRow: (id: string, file: File) => Promise<void>;
+  fillRows: (files: File[]) => Promise<void>;
+  rerunRows: (ids: string[]) => void;
+  clearRows: (ids: string[]) => void;
+  deleteRows: (ids: string[]) => void;
 
-  editCell: (
-    cell: Pick<AnswerTableCell, "rowId" | "columnId">,
-    answer: Partial<Answer>
+  editCells: (
+    cells: { rowId: string; columnId: string; cell: CellValue }[],
+    tableId?: string
   ) => void;
-  rerunCells: (cells: Pick<AnswerTableCell, "rowId" | "columnId">[]) => void;
+  rerunCells: (cells: { rowId: string; columnId: string }[]) => void;
+  clearCells: (cells: { rowId: string; columnId: string }[]) => void;
+
+  addGlobalRules: (rules: Omit<AnswerTableGlobalRule, "id">[]) => void;
+  editGlobalRule: (id: string, rule: Partial<AnswerTableGlobalRule>) => void;
+  deleteGlobalRules: (ids?: string[]) => void;
+
+  openChunks: (cells: { rowId: string; columnId: string }[]) => void;
+  closeChunks: () => void;
 
   addFilter: (filter: Omit<AnswerTableFilter, "id">) => void;
   editFilter: (id: string, filter: Partial<AnswerTableFilter>) => void;
   deleteFilters: (ids?: string[]) => void;
   applyFilters: () => void;
 
-  clear: () => void;
-
-  _syncAnswers: () => Promise<void>;
-  _upsertCell: (cell: AnswerTableCell) => void;
-  _getCellMap: () => Map<string, AnswerTableCell>;
+  clear: (allTables?: boolean) => void;
 }
 
-export type Document = z.infer<typeof documentSchema>;
-
-export interface Prompt {
+export interface AnswerTable {
   id: string;
-  entityType: string;
-  query: string;
-  type: "int" | "str" | "bool" | "int_array" | "str_array";
-  rules: PromptRule[];
+  name: string;
+  columns: AnswerTableColumn[];
+  rows: AnswerTableRow[];
+  globalRules: AnswerTableGlobalRule[];
+  filters: AnswerTableFilter[];
+  chunks: Record<CellKey, Chunk[]>;
+  openedChunks: CellKey[];
+  loadingCells: Record<CellKey, true>;
+  uploadingFiles: boolean;
 }
-
-export type Answer = z.infer<typeof answerSchema>;
-
-export type PromptRule =
-  | {
-      type: "must_return" | "may_return";
-      options: string[];
-    }
-  | {
-      type: "max_length";
-      length: number;
-    };
 
 export interface AnswerTableColumn {
   id: string;
-  prompt: Prompt;
   width: number;
   hidden: boolean;
+  entityType: string;
+  type: "int" | "str" | "bool" | "int_array" | "str_array";
+  generate: boolean;
+  query: string;
+  rules: AnswerTableRule[];
 }
 
 export interface AnswerTableRow {
   id: string;
-  document: Document;
+  sourceData: SourceData | null;
   hidden: boolean;
-}
-
-export interface AnswerTableCell {
-  columnId: string;
-  rowId: string;
-  dirty: boolean;
-  answer: Answer;
+  cells: Record<string, CellValue>;
 }
 
 export interface AnswerTableFilter {
@@ -91,3 +91,24 @@ export interface AnswerTableFilter {
   criteria: "contains" | "contains_not";
   value: string;
 }
+
+export interface AnswerTableGlobalRule extends AnswerTableRule {
+  id: string;
+  entityType: string;
+}
+
+export interface AnswerTableRule {
+  type: "must_return" | "may_return" | "max_length" | "replace";
+  options?: string[];
+  length?: number;
+}
+
+export type SourceData = {
+  type: "document";
+  document: Document;
+};
+
+export type CellKey = `${string}-${string}`;
+export type Document = z.infer<typeof documentSchema>;
+export type CellValue = z.infer<typeof answerSchema> | undefined;
+export type Chunk = z.infer<typeof chunkSchema>;